+++ src/core/js/tracking.js
@@ -71,6 +71,11 @@ class Tracking extends Backbone.Controller {
     if (completionData.status === COMPLETION_STATE.INCOMPLETE) {
       return;
     }
+    
+    const canRetry = (completionData.assessment?.canRetry === true);
+    if (completionData.status === COMPLETION_STATE.FAILED && canRetry) {
+      return;
+    }
 
     Adapt.trigger('tracking:complete', completionData);
     logging.debug('tracking:complete', completionData);
@@ -127,12 +132,4 @@ class Tracking extends Backbone.Controller {
 }
 
 const tracking = new Tracking();
-
-Object.defineProperty(Adapt, 'tracking', {
-  get() {
-    logging.deprecated('Adapt.tracking, please use core/js/tracking directly');
-    return tracking;
-  }
-});
-
 export default tracking;